+++ familyhelper-project-impl/src/test/java/com/dwarfeng/familyhelper/project/impl/configuration/CacheConfiguration.java
@@ -36,8 +36,6 @@ public class CacheConfiguration {
     private String taskPrefix;
     @Value("${cache.prefix.entity.pre_task}")
     private String preTaskPrefix;
-    @Value("${cache.prefix.entity.time_point}")
-    private String timePointPrefix;
     @Value("${cache.prefix.entity.memo}")
     private String memoPrefix;
     @Value("${cache.prefix.entity.memo_file_info}")
@@ -106,16 +104,6 @@ public class CacheConfiguration {
                 (RedisTemplate<String, FastJsonPreTask>) template,
                 new TpStringKeyFormatter(preTaskPrefix),
                 new DozerBeanTransformer<>(PreTask.class, FastJsonPreTask.class, mapper)
-        );
-    }
-
-    @Bean
-    @SuppressWarnings("unchecked")
-    public RedisBatchBaseCache<LongIdKey, TimePoint, FastJsonTimePoint> timePointRedisBatchBaseCache() {
-        return new RedisBatchBaseCache<>(
-                (RedisTemplate<String, FastJsonTimePoint>) template,
-                new LongIdStringKeyFormatter(timePointPrefix),
-                new DozerBeanTransformer<>(TimePoint.class, FastJsonTimePoint.class, mapper)
         );
     }