+++ familyhelper-project-node/src/main/java/com/dwarfeng/familyhelper/project/node/configuration/FastJsonConfiguration.java
@@ -20,7 +20,6 @@ public class FastJsonConfiguration {
         ParserConfig.getGlobalInstance().addAccept(FastJsonTaskTypeIndicator.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonTask.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonPreTask.class.getCanonicalName());
-        ParserConfig.getGlobalInstance().addAccept(FastJsonTimePoint.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonMemo.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonMemoFileInfo.class.getCanonicalName());
         LOGGER.debug("FastJson autotype 白名单配置完毕");